refactor: transaction pages
turning transactions pages one component to manage steps in transaction flow
