+++ src/pages/ListContacts/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { IconContext } from 'react-icons';
 import { FiSearch } from 'react-icons/fi';
+import { useHistory } from 'react-router-dom';
 import BalanceHeader from '../../components/molecules/BalanceHeader';
 import Header from '../../components/molecules/Header';
 import NavigationMenu from '../../components/molecules/NavigationMenu';
@@ -13,6 +14,8 @@ import {
 } from './styles';
 
 const ListContacts: React.FC = () => {
+  const { push } = useHistory();
+
   return (
     <Container>
       <Header qrCode />
@@ -24,15 +27,15 @@ const ListContacts: React.FC = () => {
         <SearchBox placeholder="Buscar contato" />
       </SearchBoxContainer>
       <ContactsContainer>
-        <ContactCard>
+        <ContactCard onClick={() => push('/transaction')}>
           <span className="name"> Jose soares costa silva</span>
           <span className="email"> josesoares@hotmail.com</span>
         </ContactCard>
-        <ContactCard>
+        <ContactCard onClick={() => push('/transaction')}>
           <span className="name"> Vitor Rafael Rocha Campelo</span>
           <span className="email"> vitinho123@gmail.com</span>
         </ContactCard>
-        <ContactCard>
+        <ContactCard onClick={() => push('/transaction')}>
           <span className="name"> Jesus Santos di Maria</span>
           <span className="email"> jj_tudo_pode321@hotmail.com</span>
         </ContactCard>
