+++ src/pages/NewTransaction/styles.ts
@@ -15,11 +15,8 @@ export const Container = styled.div`
   align-items: center;
   justify-content: space-between;
   background-color: #f7f7f7;
-  button {
-    margin-bottom: 40px;
-  }
+
   span {
-    width: 100vw;
     font-size: 18px;
     font-weight: 600;
     line-height: 27px;
@@ -53,5 +50,57 @@ export const InputContainer = styled.div<InputProps>`
     background-color: #f7f7f7;
     border: 0;
     width: 100%;
+  }
+`;
+
+export const PasswordInputContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  height: 250px;
+  a {
+    color: #9b9b9b;
+    font-size: 14px;
+    margin-top: 16px;
+  }
+`;
+
+export const ButtonContainer = styled.div`
+  width: 100vw;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  padding-bottom: 20px;
+  button {
+    margin-bottom: 0;
+  }
+`;
+
+export const Input = styled.input`
+  height: 55px;
+  width: 255px;
+  background-color: #d2d1d3;
+  border-radius: 10px;
+  border: 0;
+  font-size: 50px;
+  font-weight: 600;
+  text-align: center;
+`;
+
+export const ImageContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  img {
+    max-width: 90%;
+  }
+  span {
+    font-size: 24px;
+    font-weight: 600;
+    text-align: center;
+    line-height: 36px;
   }
 `;
